+++ App.css
@@ -1,0 +1,5 @@
+.pussy {
+  min-height: 100vh;
+  background-image: linear-gradient(to top, #a18cd1 0%, #fbc2eb 100%);
+  align-items: center;
+}
