+++ package.json
@@ -12,6 +12,7 @@
     "mongoose": "^5.11.17",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "react-native-paper": "^4.7.1",
     "react-scripts": "4.0.2",
     "web-vitals": "^1.0.1"
   },
