+++ src/index.css
@@ -1,5 +1,6 @@
-body {
+pussy {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen,
     Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
+  align-items: right;
 }
